# EGAD-Openapi-Docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /rel/{index}/select/table.{table}/columns.{columns}:
    get:
      tags:
      - relational-rest-index
      summary: "from database at index {index}, select from table table.{table} the\
        \ value in multiple columns {columns}(comma separated)"
      operationId: getMultiple
      parameters:
      - name: index
        in: path
        description: the index of the database to make a request to
        required: true
        schema:
          type: integer
          format: int32
      - name: table
        in: path
        description: sql table name
        required: true
        schema:
          type: string
      - name: columns
        in: path
        description: array of columns to get
        required: true
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: "table, value, and row exist"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /rel/select/table.{table}/columns.{columns}:
    get:
      tags:
      - relational-rest
      summary: "select from table table.{table} the value in multiple columns {columns}(comma\
        \ separated)"
      operationId: getMultiple_1
      parameters:
      - name: table
        in: path
        description: sql table name
        required: true
        schema:
          type: string
      - name: columns
        in: path
        description: array of columns to get
        required: true
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: "table, value, and row exist"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTable'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /kv/{index}/type/key.{key}:
    get:
      tags:
      - kv-rest-index
      summary: "get the type of key {key} from database in index {index}"
      operationId: getType
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringData'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /kv/{index}/getMul/key.{key}:
    get:
      tags:
      - kv-rest-index
      summary: "get multiple keys {key} (comma separated list) from database in index\
        \ {index}"
      operationId: getMultiple_2
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /kv/type/key.{key}:
    get:
      tags:
      - kv-rest
      summary: "get the type of key {key} from all databases"
      operationId: getType_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListString'
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/setex/key.{key}/value.{value}/t.{time}:
    get:
      tags:
      - kv-rest
      summary: "Set key {key} to hold the string value {value} and set key to timeout\
        \ after a given number of seconds {t} in every database"
      operationId: setEx_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: value
        in: path
        required: true
        schema:
          type: string
      - name: time
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/set/key.{key}/value.{value}:
    get:
      tags:
      - kv-rest
      summary: "set key {key} to value {value} in every database"
      operationId: setKey_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: value
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/incrby/key.{key}/incr.{incr}:
    get:
      tags:
      - kv-rest
      summary: "increment key {key} by increment amount {inc} in all databases"
      operationId: incr_2
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: incr
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLong'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/incr/key.{key}:
    get:
      tags:
      - kv-rest
      summary: "increment key {key} in all databases"
      operationId: incr_3
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLong'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/getMul/key.{key}:
    get:
      tags:
      - kv-rest
      summary: "get multiple keys {key} (comma separated list) from all databases"
      operationId: getMultiple_3
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: array
          items:
            type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTable'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/get/key.{key}:
    get:
      tags:
      - kv-rest
      summary: "get key {key} from every database"
      operationId: getKey_3
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListString'
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/expire/key.{key}/t.{time}:
    get:
      tags:
      - kv-rest
      summary: "set key {key} to expire after time {t} seconds in every database"
      operationId: expire_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: time
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/exists/key.{key}:
    get:
      tags:
      - kv-rest
      summary: "does key {key} exist in every database"
      operationId: exists_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBool'
  /kv/del/key.{key}:
    get:
      tags:
      - kv-rest
      summary: "delete key {key} in all databases"
      operationId: del_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/addMul/key.{key}/value.{value}:
    get:
      tags:
      - kv-rest
      summary: "add multiple keys {key} (comma separated list) with values {value}\
        \ (comma separated list) to all databases"
      operationId: addMul_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: array
          items:
            type: string
      - name: value
        in: path
        required: true
        schema:
          type: array
          items:
            type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        rc:
          type: string
          enum:
          - OK
          - CREATED
          - ACCEPTED
          - NO_CONTENT
          - NOT_MODIFIED
          - BAD_REQUEST
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - TOO_MANY_REQUESTS
          - SERVICE_UNAVAILABLE
        responseCode:
          type: string
          enum:
          - OK
          - CREATED
          - ACCEPTED
          - NO_CONTENT
          - NOT_MODIFIED
          - BAD_REQUEST
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - TOO_MANY_REQUESTS
          - SERVICE_UNAVAILABLE
    Table:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            type: string
    ListTable:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            additionalProperties:
              type: string
    ListString:
      type: object
      properties:
        data:
          type: array
          items:
            type: string
    StringData:
      type: object
      properties:
        data:
          type: string
...
    ListLong:
      type: object
      properties:
        data:
          type: array
          items:
            type: integer
            format: int64
    ListBool:
      type: object
      properties:
        data:
          type: array
          items:
            type: boolean
...